requirements.txt
torch
torchvision
Pillow
pandas
tqdm
open-clip-torch
flask
